Add mutiple handles
In the settings, 'dragger' no longer points to a single element's ID, it is a boolean stating whether or not there is going to be at least one handle in the page. Then, to set the element/s you want as the handle/s, just add 'data-snap-handle="true"' to their opening tag.